crazyflie_cpp: Change the thrust type from uint16_t to float.

+++ crazyflie_cpp/include/crazyflie_cpp/Crazyflie.h
@@ -73,7 +73,7 @@ public:
     float roll,
     float pitch,
     float yawrate,
-    uint16_t thrust);
+    float thrust);
 
   void sendExternalPositionUpdate(
     float x,
